requirements.txt
openpyxl==3.1.5
requests==2.32.3
fastapi==0.115.7
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
